Refactor IO operation result types for consistency and clarity
- Updated the return types of various IO operations from `io::Result<u32>` to `io::Result<usize>` to ensure consistency across the codebase.
- Modified related functions and traits, including `Driver`, `OpLifecycle`, and network operations, to reflect the new result type.
- Adjusted tests to validate the changes in return types, ensuring correctness in asynchronous operations.

+++ veloq-runtime/src/io/driver.rs
@@ -17,7 +17,7 @@ pub trait Driver {
         &mut self,
         user_data: usize,
         cx: &mut Context<'_>,
-    ) -> Poll<(io::Result<u32>, IoResources)>;
+    ) -> Poll<(io::Result<usize>, IoResources)>;
 
     /// Submit queued operations to the kernel.
     fn submit(&mut self) -> io::Result<()>;
