+++ veloq-runtime/src/io/op/linux.rs
@@ -50,7 +50,7 @@ impl OpLifecycle for Accept {
         }
     }
 
-    fn into_output(self, res: std::io::Result<u32>) -> std::io::Result<Self::Output> {
+    fn into_output(self, res: std::io::Result<usize>) -> std::io::Result<Self::Output> {
         let fd = res? as SysRawOp;
         use crate::io::socket::to_socket_addr;
         let addr = if let Some(a) = self.remote_addr {
